Parse the project url from upstream, not origin.
origin is expected to point at a fork, and upstream is expected to
point at the receiver of pull requests. In case of a maintainer only
checkout, the only remote will be the upstream remote, and origin
might not exist.

The actual project name should be parsed from the upstream URL. This
will still break if the fork has renamed their version, but solving
that requires actual GitHub API calls.

Minor README text change.