Add ignore_collaborators class variable
Even if you don't have the collaborators.yaml file available, it would
still be useful to be able to run the rest of the tests. To this end,
add a class variable "ignore_collaborators" which, when set, skips
the loading of the collaborators.yaml file and accepts all signers.
The GnuPG signatures still need to validate.